reorganized folder structure to meet react standards
